Added support for timeStep
git-svn-id: http://swe-common-data-framework.googlecode.com/svn/common@34 4a6684ca-5c49-0410-8270-e1cb95179ea8

+++ src/org/vast/util/DateTimeFormat.java
@@ -224,7 +224,7 @@ public class DateTimeFormat extends SimpleDateFormat
      */
     static public String formatIsoPeriod(double periodSeconds)
     {
-        return "PT" + periodSeconds + "S";
+        return "PT" + (int)periodSeconds + "S";
     }
